libft.h and other changes

+++ libft/ft_bzero.c
@@ -1,6 +1,8 @@
-void ft_bzero(void *s, unsigned int n)
+#include "libft.h"
+
+void ft_bzero(void *s, size_t n)
 {
-    unsigned int i = 0;
+    size_t i = 0;
 
     while (i < n)
     {
