+++ libft/ft_memset.c
@@ -1,6 +1,8 @@
-void *ft_memset(void *s, int c, unsigned int n)
+#include "libft.h"
+
+void *ft_memset(void *s, int c, size_t n)
 {
-    unsigned int i = 0;
+    size_t i = 0;
 
     while (i < n)
     {
