+++ libft/ft_strlcpy.c
@@ -1,7 +1,9 @@
-unsigned int	ft_strlcpy(char *dest, const char *src, unsigned int size)
+#include "libft.h"
+
+size_t	ft_strlcpy(char *dest, const char *src, size_t size)
 {
-	unsigned int	i;
-	unsigned int	res;
+	size_t	i;
+	size_t	res;
 
 	i = 0;
 	res = 0;
